Use elaborated form of cases to check for defaults
Before, the check for a default catch-all case was done using PTerm and
the Placeholder constructor. Now, universal cases at the end are found
by checking that all arguments are elaborated to pattern variables. This
can catch cases where it's not an explicit placeholder, and it avoids
confusion with constructors by checking the binders of the elaborated
case.