Added whole R/W effectful file ops.
Operations on success will return the contents of the file (for read),
and Unit for write.  Upon failure the functions will return a user
supplied data type that is used to store the 'file name' upon which
these operations were called. This is to allow the user to decide if
these operations should raise exceptions, and also how to represent
the error.

Further, default forms have been supplied that 'just' return the file
name upon failure.