Minor change to display of metavariables
Displaying things with machine generated names is really just noise, and
potentially confusing since they're inaccessible and typically an
artefact of the elaboration process, so hide them.

(If this turns out to be differently confusing, we'll have to consider
an option to display them. Maybe we should do that anyway.)