Allow certain package options to be overridden from console.
This PR addresses in part: #1448, #499.

When building packages a preapproved command line options (if
specified) will be merged into the options list generated from the
iPKG file. The allowed options are placed at the head of the options
list such that they. The allowed options are:

+ Logging levels
+ Default total, default partial
+ Warn partial
+ Warn reach
+ Location of IBC subdir
+ Location of import dir.

Idris also reports an error if invalid command lines options are
specified when used to override package specified options.

The code will also facilitate specific reporting of the offending
flags, if we choose to want such functionality.