Be smarter about running travis and stack.
This change optimises the travis build script for building stack. The
Travis setup has been changed to include a build job purely for
building with stack. To ensure that the stack test is isolated, and to
reduce duplication of the act of building idris, the test script has
required an additional env variable.