Friendlier error on proof search failure
Now, proof search doesn't throw an INTERNAL ERROR when called on a goal
that isn't a datatype. Instead, it provides the following message:

    Prover error: Attempted to find an element of type Nat -> Bool using
    proof search, but proof search only works on datatypes with
    constructors.

When the goal is a function type, which seems to be a likely reason why
it would fail, it adds the sentence:

    In particular, function types are not supported.

Additionally, a variant of the Msg constructor for errors, called
FancyMsg, was introduced. This allows the use of the ErrorReportPart
type to define custom messages.

Fixes #2713.