PParser.hs: Add |filename| parser
As it stands, it reads in either a (possibly namespaced) identifier,
which was the old behavior, or a double-quoted string, which is the new
behavior.

This could be improved by actually testing that it's returning a filename,
but I'm clumsy with the parser abstraction, and it wasn't clear to me
from the Trifecta docs I looked through how to pass it through
System.FilePath.takeFileName and then enforce that it was a non-empty string.