Migrated travis scripts to external file.
The bash scripts defined in a travis configuration file cannot be as
complicated as you can make in a standard bash script. Parsing of
complicated Bash scripts by Travis has been known to cause problems
during the build process [1].  This commit migrates several of the
more complicated scripts to a new folder `scripts`.

[1] https://github.com/travis-ci/travis-ci/issues/494

 complicated scripts to external files