Take care when calling allNamesIn
If we've expanded alternatives, the term could be very big, and we'll
get the same results if we call before expanding alternatives. If we do
need to know all the names a user has written in a term, call before
expanding the alternatives.

This addresses part of the slowness identified in #2657