Tweak desugaring rules for 'case' in do blocks
(Also for ?holes) If there's no binding, assume that the type being
returned by the case block or metavariable is ().

This makes it possible to use case and holes in the middle of do blocks
without any ugly hacks, and if there's no binding it's probably what was
intended anyway.