Bug fix for infinite loop while executing a callback in the Javascript runtime.
The NULL BC operation, which is supposed to clear a register (according to the docs) was translated as delete in the Javascript backend. This worked fine for clearing registers from the stack, but did not work well while trying to clear i$RT. This cause an infinite loop in i$ffiWrap which was checking the return value should have been cleared but continued to hold its old value (essentially ignoring the delete operation). Callbacks work fine now in my tests.