Initial commit for a category based logger for Idris.
Currently logging is based on horizontally defined levels across the
entire language infrastructure. This commit introduces vertical splits.

Logging statements record both the level the message is to be recorded
at and the categories it belongs too. No categories is an alias for
belonging to all categories.

The State record stores both the current logging level, and also the
categories to show.

Logging is decided if a given logging statement is at the right level
and if the categories specified are listed. By default all categories
are allowed, as levels will take care of logging statement visibility.