Make test script stack aware.
This commit adds some documentation to the test script file, and more
importantly makes the script stack aware.

When running a test, the script already checks to see how idris is
being built if a cabal sandbox is not being used an additional check
for stack is now made. If neither of the tools were used to build
idris then it is assumed that the default set up is used.