First attempt at getting order-independent Effects working.
This implementation has what seems to me to be unnecessary complexity
due to needing to figure out how to get the Handler eff m type class
resolution to be inferred through-out the Env re-ordering code - envElem
and dropEnv is a good example of this. I think that possibly someone else
could make this a bit cleaner, but it works in my tests.