Improve C FFI
Implement peek and poke for pointers

Move peek, poke for bitstrings to idris_bitrsing

Add access to global variables

Add call of foreign functions via pointer

This adds some new features to the C FFI that match what's in the Haskell FFI.

Prefixing the name with `&` gives a pointer to a C global.

```
errno : IO Ptr
errno = foreign FFI_C "&__errno" (IO Ptr)
```

Putting `%wrapper` in the name argument gives a function pointer to a
wrapped Idris function (useful for APIs that take a struct with a bunch
of functions). It's functionality matches "wrapper" in Haskell, but the
illegal C name makes name clashes go away.

Putting `%dynamic` in the name argument calls a C function from a
function pointer. This is useful for APIs that provide structures with
C functions (popular in object oriented C) or otherwise returns C function
pointers. es in a declaration. The type signature in dynamic should take
an extra Ptr up front for the function pointer.