Fix behaviour of %hide
%hide now merely makes a name inaccessible to programs, without changing
any further behaviour. This is by adding a new accessibility class
"Hidden" which can only be set via a %hide directive, and only for names
which are currently visible.

(Note that this doesn't quite fix the behaviour completely, because
it'll behave oddly for names hidden in the current module. Will fix in a
future commit...)

Fixes #2945