Make coverage checking more conservative
If an argument is a non-concrete type (i.e. a function type which
calculates its type from other arguments) then we can't merely generate
all the possible patterns from existing patterns, since there could be
any number of constructors of types we haven't thought of.

We now take the conservative approach of checking the catch-all pattern for
such cases, so that coverage depends entirely on the other concrete
patterns being complete.

There is likely a more precise way to do this, but this more
conservative way is at least sound.

Fixes #1292