Fixes #499 allowing packages to be installed to custom location.
When installing idris packages, specifying the command option
`--ibcsubdir <dir>` will install the package to `<dir>`.

This PR is to provide an option complimenting the env `TARGET` as the
option `-i <path>` compliments `IDRIS_LIBRARY_PATH`.

This option doesn't affect the other package commands, as they should
be performed locale to the package source. Aside from test that goes
to a temp directory.