Extended syntax for 'with' blocks
EXPERIMENTAL: I'm not sure this is a good idea or how it will turn out
to affect some other things, but let's see how it works out.

This allow calling functions defined by 'with' directly (i.e., the
generated auxiliary function) by allowing the additional argument syntax
on the right hand side. For example:

data SnocList : List a -> Type where
     Empty : SnocList []
     Snoc : SnocList xs -> SnocList (xs ++ [x])

snocList : (xs : List a) -> SnocList a

my_reverse : List a -> List a
my_reverse xs with (snocList xs)
  my_reverse [] | Empty = []
    my_reverse (ys ++ [x]) | (Snoc p) = x :: my_reverse ys | p

The '| p' on the right hand side means that the with block function will
be called directly, so the recursive structure of SnocList can direct
the recursion structure of my_reverse.