Generalise help messages for Hardware Specific Options.
The options `--target TRIPLE` and `--cpu CPU` are only really used by
the llvm-codegen, which has been spun out from the main idris project
for some time. This PR generalises and help message for these options
to be non-llvm specific, and makes reference that the backend might
not use this information during compilation.