Fixes #3075 by adding generalised 'Project' output type.
The code generator output type `MavenProject` was specific to the Java
codegen. This output type has been generalised to `Project` such that
if a codegen wants to output a 'project' it can do so. An activation
flag `--projectonly` has been added. The existing `--mvn` flag still
exists for backwards compatibility, and will be deprecated in a future
release.