New implementation of rewrite
Not using the 'rewrite' tactic any more, but building an application of
a rewriting lemma with a predicate directly. This allows a new syntax:

rewrite [rule] using [rewrite_lemma] in [scope]

Where rewrite_lemma is any function of the right shape, so can be a
function used to rewrite dependent types rather than the default, which
is rewrite__impl.

More to do: the rewrite_lemma could be generated automatically for many
dependent types, also the predicate building needs a bit of polish, but
this already works for, e.g., rewriting vector functions.