Use bit-pattern identity for Double
Change the conversion checker to use bit-pattern identity on Double
instead of the IEEE comparison that Haskell's Eq type class implements.
Previously, we had definitionally equal values like -0.0 and 0.0 that
could nevertheless be distinguished by programs. This breaking of
functionality allowed proving a contradiction.

Thanks to Ryan Scott for the tip on how to do the bit-pattern
comparison!

Fixes #2609.