Fixes #2049
These changes augments the build setup to check the sphinx documentation.

The changes are based upon the configuration used by [PyMSQL
project](https://github.com/pymssql/pymssql).

A problem with this approach is that the matrix for checking the docs
requires Idris to be built. The travis build could be made more modal
so that only the sphinx docs are built.