Lazy elaboration tweaks
Firstly, never add a Delay to a pattern variable on the lhs, since
evaluation isn't forced until it is deconstructed.
Secondly, "try" is supposed to return the first error if there's a
failure, but this wasn't happening in some cases. In particular, this
can make errors from Lazy elaboration surprising, so this is fixed.

Fixes #3105