Totality checking now looks inside case directly
This is instead of treating case blocks as mutually defined functions.
This gives better results (strictly more things are spotted as total)
because it means we don't lose information from the outer scope.

Also fixed a couple of errors on the way, of things which should
have been total but reported otherwise.