Allow 'data' declarations in interfaces
This leads to a method and a top level function as normal, but
an implementation of the interface must define the method by returing
a type constructor (or some other injective function application).

The effect is that Idris treats the top level function as a data type,
which helps with unification and interface resolution.