Set a limit on partial evaluation depth
There's reasonable situations where it could go forever if unchecked,
such as interfaces on nested types, where it won't stop because it
needs to specialise for each level of nesting. This limit stops it
elaborating forever.

Ideally it would also be looking for whether it had partially evaluated
the specific function - this limit may be a bit limiting for partially
evaluated DSLs.

Fixes #3191