Fix ordering of dependent patterns/implicits
When we reorder automatically bound variables such as implicits and
pattern variables, we need to keep explicitly bound things such as pi in
the same order as they were given by the programmer. This rewrites
orderPats to do the job properly, and moves it to the elaborator
(Elab.Utils) where it belongs (out of Core.TT).

Fixes #1944