Deprecate '%assert_total'
This is for two reasons:

1. It is error prone, because if you edit the function, the reason for
assertion may no longer be valid, but it's no longer checked. Better to
be precise about why a function is asserted total using a combination of
assert_total, assert_smaller and (new) assert_unreachable

2. It is a very tedious maintenance burden.

Reason 1 is the good reason. Reason 2 is why I actually did this.