Allow CDATA to be copied to other VMs
When using `fork` with access to a CData reference, the RTS will panic
on the assertion that it was unable to match the type of the closure in
doCopyTo. This is because of a missing case for CData.

This change copies the paradigm for other type tags of Closures for
CData.