Extract exit status with WIFEXITED/WEXITSTATUS.
It seems that extracting the exit status is more complicated than
simply looking at the `int` returned from `system`. I've wrapped
`WIFEXITED` and `WEXITSTATUS` in FFI_C in the test case to show that
it can be done.

- See http://pubs.opengroup.org/onlinepubs/9699919799/functions/system.html
- and http://pubs.opengroup.org/onlinepubs/9699919799/functions/waitpid.html

These foreign functions could be added to `libs/base/System.idr` but I'm unsure
of their portability (particularly on Windows platforms).