Remove incorrect CPP and require newer Cabal instead
The `MIN_VERSION_*` macros are only generated during setup, i.e. after Setup.hs
is compiled. The GHC version is not a proxy for the Cabal version, as it is
trivial to install a newer version of Cabal and cabal-install with
cabal-install, and in that case the Setup.hs would not typecheck. Since it
doesn’t appear to be possible to switch on the Cabal version, this change
simply declares a requirement for a newer version.