When asked Idris will 'audit' Idris projects.
It would be nice to be able to audit Idris packages for 'common' mistakes, and warn developers of *potential* problems.
This commit introduces a means to 'audit' idris files and warn when asked.

These checks are called with the flag `--warnipkg`.

Currently there is only one check: Idris files specified in the package's directory but not in the ipkg file itself.