Overidable overlapping implementation check
Using %overlapping flag on the implementation
In theory, this should never be necessary, but in practice there are
situations where Idris might think implementations are overlapping,
e.g. of the form "Foo a => Bar a", where there are other implementations
of Bar without the superclass constraint.

So, the check needs to be better, but it's a fairly big task so this is
a workaround for the moment.