Updating range behavior to match Haskell's where appropriate
This resolves #3502 and resolves #3503. This also fixes the issue that natEnumFromThenTo could not descend.

I used make && make test on Ubuntu.

After adding the new tests, the existing implementation crashed.

    basic013:
Error: test/basic013
run: line 3: 19432 Segmentation fault      (core dumped) ./basic013
With the new implementation, all 202 tests passed.

I also built this on Windows, but was unable to run the tests, so I generated a stress test suite that passed range definitions to both Idris and Haskell to verify the output would be equivalent.

There's one notable difference from Haskell. In Haskell, ranges of the form [A,A..B] generate an infinite sequence of As. I chose to follow the existing implementation, which returns [] instead. This rule conflicts with another rule that ranges of the the form [A,B..A] return [A] in the case that B == A (i.e. the form [A,A..A]). My implementation returns [A] in that case.