Fix the warnings output by cabal-install after raising the cabal-version
Specifically, if the `cabal-version` constraint is above 1.10, each component
should have a `default-language` field. It appears the default was/is
`Haskell98`, so I entered that for each component.

Similarly, the `extensions` field is then deprecated in favor of
`default-extensions` and `other-extensions`. For these I have investigated which
extensions are actually required globally (i.e. some files assume them but do
not declare them) and have entered these as `default-extensions` and removed
them from any files that did declare them. I then collected all additional
extensions declared in any files into `other-extensions`.

It may be a good idea to think about these settings and change them, and also to
prune declared but unused extensions from the source files, but in this commit I
have only tried to replicate the previous settings.