Change totality checking and inlining
This solves a problem caused by inlining for totality checking, but not
elsewhere. For example, the following passes totality checking because x
is inlined:

foo = let x = diverging in 42

If we then assume it's safe to evaluate all parts of foo because it's
total, as the partial evaluator does, the compiler ends up in a loop.
So, the totality checked version and the stored version really need to
be the same, which this patch does.

We still need to do some controlled unfolding so that interface
dictionaries work correctly (controlled using the 'Unfold' flag on a
function), and to spot definitions which preserve
guardedness of productive definitions.