Fix a source of major slowness
When we partially evaluate a definition, we need to know which names to
unfold (those that are part of the definition) so that we don't unfold
too much of the arguments in static position. We do this by searching
the call graph, which is really slow if we don't cache it. So, when we
elaborate a function with static arguments, now we cache it.