Apply forcing optimisation on LHS
This means that the case trees generated are much smaller, because we
don't bother matching on things we already know about.

Needs some care, because it affects some existing erasure results. For
the moment, we only force arguments when there's *no* occurrence of a
variable below a non-constructor name. This may be an issue with erasure
warnings rather than forcing, but it solves the problem for now.

Fixes #2657