New coverage checker
This now works by building a case tree and generating the missing
clauses from that, with some trickery to deal with overlapping cases.
The result is much faster, and fixes some existing (unreported...)
issues.

I have exercised this quite a bit, and added a couple of new test cases
while I'm at it, but it is an undecidable problem so it's not impossible
that some existing programs will no longer compile, even if that is
unlikely. If such programs aren't fixable with extra "impossible" case
annotations, it may be a bug...