Updated CONTRIBUTORS to reflect actual contributors.
Sadly we haven't be good at keeping the list of contributors up-to-date.

This fix takes the unique list fo contributors from the git logs, and
updated the list in CONTRIBUTORS. There may be a duplication of contributors
whose usernames couldn't be resolved to proper names. Duplicate names were
remove in the first place.