Fix verbosity configuration error.
There was an existing 'logical' flaw for when to report progress.
When using a custom codegen (i.e. non-`C`), Idris would be verbose during
compilation even if no verbosity flag was set. For the default codegen
everything was non-verbose as expected: Type checking and parsing were not
verbosely reported.

The fix was to set the default verbosity to False. I suspect the error was
a result of the default codegen correctly setting verbosity to false, and
other codegen's weren't.