Change evaluation of partial functions at REPL
To prevent infinite recursion at the REPL under binders, which results
in a neverending display of a diverging term, the evaluator at the REPL
will now block on a partial function when there is a free variable in
the context.

This means that you can still test partial functions at the REPL with
canonical values, but if they are recursive and non-terminating under
binders, you won't get infinite evaluation.

Fixes #3535