Fixed the Num instance for Complex
Due to a historical artifact, the `Num` implementation for `Complex` only covered `Double`, because
`abs` used to be in `Num` (like in Haskell). Since these implementations are now split, we can
implent Num with only `Neg a` instead of Double.

As soon as `abs` moves out of `Neg` (issue #3500), we can also loosen the requirment for the `Neg`
implementation to only need `Neg a`:

```idris
implementation Neg a => Neg (Complex a) where
    negate = map negate
    (-) (a:+b) (c:+d) = ((a-c):+(b-d))
```