Fix patvar naming for implicits
There's no need to make the name unique if the pattern is going to have
the same name as the hole. Otherwise, this can cause problems under the
types of metavariables, where the name gets renamed and the generated
type is then no longer valid because there's an undefined name in it.
Fixes #3398