Fix for cyclic solutions in unifyProblems
If we get x=y and y=x, need to update the second from the result of the
first so that we don't solve holes that aren't actually solved yet. This
fixes one cause of the mysterious "No such variable" error by not
solving a hole with a variable that doesn't exist.
Fixes #2731