Moved initialization code to Main.hs
There is a bug in stack where the config hook of Setup.hs is not being
fired when a flag is changed just in stack.yaml. So, in order to prevent
others from wasting their time with this, the init code has moved to
Main.hs, where it is slightly more inconvenient to change.

Also made small cosmetic cleanups.