Bring back external codegens
In a previous commit I removed the support for the external codegens.
But later on the refactoring project I decided to make it conservative.
So I am bringing it all back. Note that the Codegen type still changed,
so the IBC version is still updated.
I also added 2 TODOs where I noticed previously incomplete implementations.