A simple extension system for code generators
To use this: create a package that depends on `idris` and call the
provided register* functions in the your `main` to register your codegen.
You can add extra lib/include flags and info strings that will be displayed
by `idris --info`. The result will be an executable that has your codegen
embedded and available in the REPL or IDE mode. Using it will not incur
the overhead associated with loading IBC files from Idris prelude or other
dependencies that may have been already loaded.