Check for node when running test suite.
If node isn't installed then don't run the test suite. This check
could possibly be performed when the test option for running node
is set, but it is not clear how.

Regardless, if you are contributing a pull request then you would
need to run the tests for both `C` and `Node` codegens.