Added Control.ST to contrib
This is a library which allows composition of state transitions.
Documentation and a paper will be available soon...

This does essentially what effects does, but with easier creation and
destruction of resources, the ability to implement interfaces in terms
of other effectful interfaces, neater description of types, and some
error reflection.