Add Control.ST.Exception
Only possible to throw exceptions after all the state transitions have
been done, meaning that we always know everything is in the final state
when we get to the catch block.

This is most likely to be useful for controlling things, such as
backtracking parsers, so it's possible that 'Exception' is not really
the right name for it. However...