Fix for implicits in records
Since we have this annoying distinction between top level implicits and
scoped implicits, we need to make sure that things are marked
appropriately. So when we parse record fields, they become scoped
implicits.

(I don't think there is really a need for this distinction, but because
implicits were top level only at first, it's hard to unify everything.
One for the next rewrite, should it ever happen, I suppose...)

Fixes #2266