Take some more care with forcing
If the forced name is only forceable because we've reduced a name, and
that name isn't exported, then we can't actually get the name in other
modules! So, check for forceable names before reducing the type, not
after.

(We could do a bit better than this, in fact, and reduce names which are
public export, but there isn't a huge benefit to this at the moment so
I'll save it for later.)