Deal with ambiguity in impossible clauses
Some things were getting past totality checking because we weren't
generating clauses when there was ambiguity on a non-checkable lhs. Now
take a type-directed approach to building the term for the case tree
which, even though it can't be checked, at least has the right
unambiguous names.

If we can't resolve the ambiguity, it's an error. In some (probably
rare) cases this might need the programmer to disambiguate explicitly. I
think this is a price worth paying for soundness...

Fixes #3269