Change meaning of 'data' in interfaces
A "data" declaration in an interface means "this must be implemented by
something injective, so that unification can determine the arguments".

This is slightly weaker than the existing definition, which required the
definition to be an actual data type. It means that unification can
learn less, but it is far more useful in practice, it turns out.

This breaks the interfaces005 test, but it is not clear that this test
is testing something that we want to do, and if we do want to do it, we
can do it a different way...