Handle idris_mkFileError through a special-case.
Use a dirty IORef to simulate C's errno behaviour. If an exception
occurs during fileOpen, then store a FileError expression, which is read
and returned in a later call to idris_mkFileError.

As a result of this change, file errors generated during the execution
of a type provider, for example file not found, will be correctly
reported. Previously the interpreter crashed.

Fixes #3716.