Support presentations of abstract names in output
When Idris sends strings over the IDE protocol, it includes
out-of-band semantic information that allows regions of the string to
_represent_ their underlying data. In particular, terms are serialized
and sent with their representations, allowing things like interactive
error messages in IDE clients.

This patch adds a similar capability for names. Now, serialized
datatypes are sent with names, allowing names that don't have syntax
to be the subjects of commands. This is very convenient when working
with metaprogramming or debugging the compiler.

The change is backwards-compatible, because it has to be enabled in
each IDE client, and if they aren't sent serialized names, then they
don't act differently from today.