Addition of a literate tooling to `contrib`.
`Text.Literate` is a simple Idris module to help with processing
literate document, and returning the raw code.

The tooling is markup language and code agnostic, and uses
user-defined 'styles' to identify code blocks in the presented text
document. This allows literate styles to be constructed that respect
the host markup language's conventions.

The underlying tokeniser is also greedy. Once it identifies a line
marker it reads until the end of line, and once it recognises a start
block, it will consume until the next identifiable end block. Thus the
input literate files must be well-formed w.r.t to line markers and
code blocks.