Fix line number reporting for `Text.Literate`.
Ensures that the line count of the original document and resulting
code file are kept the same. This is important for reporting errors
from the code in the original document. Unfortunatly, column numbers
for line markers are not preserved and differ by one plus the number
of characters in the used line marker.