Text.Lexer: add surround, quote, and escape
`surround` and `quote` are both for sequences of lexers between
existing lexers. `surround` is for when the lexers at the start and end
are different, and `quote` is for when they are the same.

`escape` recognises any single escape character followed by a sub-lexer.

Existing combinators were written in terms of these new ones where it
made sense.