Rework commit in Text.Parser
The previous commit/alternative behaviour seemed to be
trying to scope commits to a particular alternative,
rather than the parse as a whole, but it wasn't clear
to me how this was meant to work:

For example, the parser:

 Commit

would commit to the current alternative, while the
parser:

 (Commit <|> fail "X")

would not commit - seemingly because it was attempting
to commit only to the <|> used inside the parser
definition.

This violates a rule that I was expecting, that

  x <|> fail "X"

would behave the same, commitment-wise, as simply

  x

This reworked code never removes a commitment. When a
parser commits, it commits the entire parse tree.

In this model, I think it would make sense to have
a Try primitive that that removes commitment of a
failing sub-parser (see 'try' in Parsec).

I claim that whether a parser commits or not depends
only on that parser, not any parser evaluated before.

All simple parsers except Commit do not commit.

Parser combinators decide their commitness as functions
of the parameter sub-parsers, without reference to
anything else.